pyproject.toml
[project]
name = "gradio-chatbot-openai"
version = "0.1.0"
description = "A web-based AI chatbot using Gradio and OpenAI SDK"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = []
